update example tests
