+++ examples/api/test/material/selection_area/selection_area.1_test.dart
@@ -7,18 +7,15 @@ import 'package:flutter_api_samples/material/selection_area/selection_area.1.dar
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  testWidgets('SelectionArea Color Text Red Example Smoke Test', (WidgetTester tester) async {
+  testWidgets('SelectionArea SelectionListener Example Smoke Test', (WidgetTester tester) async {
     await tester.pumpWidget(
-      const example.SelectionAreaColorTextRedExampleApp(),
+      const example.SelectionAreaSelectionListenerExampleApp(),
     );
-    expect(find.widgetWithIcon(FloatingActionButton, Icons.undo), findsOneWidget);
     expect(find.byType(Column), findsNWidgets(2));
-    expect(find.textContaining('This is some bulleted list:\n'), findsOneWidget);
-    for (int i = 1; i <= 7; i += 1) {
-      expect(find.widgetWithText(Text, '• Bullet $i'), findsOneWidget);
-    }
-    expect(find.textContaining('This is some text in a text widget.'), findsOneWidget);
-    expect(find.textContaining(' This is some more text in the same text widget.'), findsOneWidget);
-    expect(find.textContaining('This is some text in another text widget.'), findsOneWidget);
+    expect(find.textContaining('Selection StartOffset:'), findsOneWidget);
+    expect(find.textContaining('Selection EndOffset:'), findsOneWidget);
+    expect(find.textContaining('Selection Status:'), findsOneWidget);
+    expect(find.textContaining('Selection Listener Status:'), findsOneWidget);
+    expect(find.textContaining('This is some text under a SelectionArea that can be selected.'), findsOneWidget);
   });
 }
